# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: API Resource for Catalog Service
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /api/v1.0
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /categories:
    x-swagger-router-controller: category
    post:
      tags: [category]
      summary: Add a new root category.
      description: Add a new root category.
      operationId: addCategory
      parameters:
        - name: restaurantId
          in: header
          description: restaurantId will be redirected from api-gateway.
          required: true
          type: string
        - name: Category
          in: body
          description: The content of the new category.
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CategorySuccessResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/BadRequestResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /categories/{parentId}/subCategories:
    x-swagger-router-controller: category
    post:
      tags: [category]
      summary: Create a sub category.
      description: Create a sub category.
      operationId: addSubCategory
      parameters:
        - name: restaurantId
          in: header
          description: restaurantId will be redirected from api-gateway.
          required: true
          type: string
        - name: parentId
          in: path
          description: The parentId which will be a parent category of the new sub category.
          required: true
          type: string
        - name: Category
          in: body
          description: The content of the new category.
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CategorySuccessResponse"

        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/BadRequestResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /categories/{categoryId}/move:
    x-swagger-router-controller: category
    post:
      tags: [category]
      summary: Move a category underneath the other one.
      description: Move a category underneath the other one.
      operationId: moveCategory
      parameters:
        - name: restaurantId
          in: header
          description: restaurantId will be redirected from api-gateway.
          required: true
          type: string
        - name: categoryId
          in: path
          description: The categoryId which will be move to to a parent category. Make it as a root one by giving parentId with ''.
          required: true
          type: string
        - name: MoveToCategory
          in: body
          description: The parentId that this categoryId will be underneath. Giving empty string will make this one become a root categoty.
          schema:
            $ref: "#/definitions/MoveToCategory"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CategorySuccessResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/BadRequestResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /categories/{categoryId}:
    x-swagger-router-controller: category
    delete:
      tags: [category]
      summary: Delete a category, but not delete its items.
      description: Delete a category, but not delete its items.
      operationId: deleteCategory
      parameters:
        - name: restaurantId
          in: header
          description: restaurantId will be redirected from api-gateway.
          required: true
          type: string
        - name: categoryId
          in: path
          description: The categoryId which will be deleted.
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CategorySuccessResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/BadRequestResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags: [category]
      summary: Update info of a category.
      description: Update info of a category. This API endpoint will not update the parentId of the category. User /categoty/{categoryId}/move instead.
      operationId: updateCategory
      parameters:
        - name: restaurantId
          in: header
          description: restaurantId will be redirected from api-gateway.
          required: true
          type: string
        - name: categoryId
          in: path
          description: The categoryId which will be deleted.
          required: true
          type: string
        - name: Category
          in: body
          description: The content of a new category.
          required: true
          schema:
            $ref: "#/definitions/UpdateCategory"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CategorySuccessResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/BadRequestResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /categoriess:
    x-swagger-router-controller: category-new
    post:
      tags: [category]
      summary: Add a new root category.
      description: Add a new root category.
      operationId: addCategory1
      parameters:
        - name: Category
          in: body
          description: The content of the new category.
          required: true
          schema:
            $ref: "#/definitions/Category-new"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CategorySuccessResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/BadRequestResponse"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/BadRequestResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
# complex objects have schema definitions
definitions:
...
  MoveToCategory:
    required:
      - parentId
    properties:
      parentId:
        description: the parentId which the category will be belong.
        type: string
  Category:
    required:
      - name
      - description
    properties:
      name:
        type: string
      description:
        type: string
      serviceFee:
        type: number
      isPlusPlusSign:
        type: boolean
      taxName:
        type: string
      taxFee:
        type: number
      language:
        description: Inputing language. Default is English.
        type: string
      translation:
        description: Array of translation for the category.
        type: array
        items:
            $ref: '#/definitions/Translation'
      assets:
        description: Including a thumbnail and an images list.
        properties:
          thumbnail:
            $ref: "#/definitions/Asset"
          images:
            type: array
            items:
              $ref: "#/definitions/Asset"
  UpdateCategory:
    properties:
      name:
        type: string
      description:
        type: string
      language:
        description: Inputing language. Default is English.
        type: string
      translation:
        description: Array of translation for the category.
        type: array
        items:
            $ref: '#/definitions/Translation'
      assets:
        description: Including a thumbnail and an images list.
        properties:
          thumbnail:
            $ref: "#/definitions/Asset"
          images:
            type: array
            items:
              $ref: "#/definitions/Asset"
  Asset:
    properties:
      source:
        type: string
      awsKey:
        type: string
  Translation:
    properties:
      fieldName:
        description: The field name will be translated. E.g. the "description"
        type: string
      content:
        description: The content of the fieldName. E.g. "Sách tham khảo" in Vietnamese.
        type: string
      languageCode:
        description: The language code being used for filtering on response. E.g. 'vi'.
        type: string
      language:
        description: The language supported for full-text search. E.g. 'finnish'. Default is 'none'.
        type: string
    required:
      - fieldName
      - content
      - languageCode
  Category-new:
    required:
      - name
      - description
    properties:
      name:
        type: string
      description:
        type: string
      logoImage:
        description: Including a thumbnail and an images list.
        properties:
          thumbnail:
            $ref: "#/definitions/Logo-image"
  Logo-image:
    properties:
      name:
        type: string
      source: 
        type: string
...
  CategorySuccessResponse:
    required:
      - categoryId
      - message
    properties:
      categoryId:
        type: string
      message:
        type: string
  BadRequestResponse:
    required:
      - code
      - message
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string